Namespace galsfunctions defined
The parameter setting file has been significantly cleaned up. It requires a little more fine tuning and then we can essentially get rid of it.
The master level set matrices and the arrays that store the node location have been initialized in a vector form now.
All the initialization is done from Allocation.h file, which consists of specific functions.
The functions have been defined in a namespace galsfunctions.