Initialization on Device itself
Instead of initializing in serial and then copying over the levelset data, now
computing the levelset data on the device itself. This saves the copying time
and computing the initial levelset data in parallel should be faster anyway.

Currently, we are calling the initialize function on the host as well, this is
only because we need to write the initial data to an output file. We can get rid
of this function call.